拆分warehouse
