+++ .gitignore
@@ -1,4 +1,3 @@
-
 build
 bin
 .DS_Store
