+++ level1/p12_warehouse/CMakeLists.txt
@@ -1,0 +1,7 @@
+cmake_minimum_required(VERSION 3.7)
+project(warehouse)
+aux_source_directory(${PROJECT_SOURCE_DIR}/src DIR_SRC)
+link_directories(${PROJECT_SOURCE_DIR}/lib)
+include_directories(${PROJECT_SOURCE_DIR}/include)
+set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
+add_executable(warehouse ${DIR_SRC})
