+++ level1/p12_warehouse/src/warehouse.c
@@ -1,58 +1,5 @@
-#include<stdio.h>
-#include<stdlib.h>
-#include<string.h>
-#include<stdbool.h>
-
-#define MAXLEN 10
-#define FILENAME "ware.txt"
-
-struct item{ 
-    char name[MAXLEN+1];
-    unsigned number;
-    struct item* next;
-};
-
-bool loadList();
-bool dumpList();
-void showList();
-struct item* findItem(char name[]);
-bool insertItem(char name[],unsigned number);
-bool deleteItem(char name[],unsigned number);
+#include"warehouse.h"
 struct item* head=NULL;
-void readMe();
-inline FILE* fOpen(char * filename,char * mode);
-
-int main(){ 
-    loadList();
-    readMe();
-
-    unsigned temp;
-    char mode,buf[MAXLEN+1];
-    while(~scanf("%c",&mode)){ 
-        mode=tolower(mode);
-        switch(mode){ 
-            case 's':
-                showList();
-                break;
-            case 'i':
-                scanf("%s %d",buf,&temp);
-                insertItem(buf,temp);
-                break;
-            case 'd':
-                scanf("%s %d",buf,&temp);
-                deleteItem(buf,temp);
-                break;
-            case 'q':
-            	dumpList();
-                return 0;
-            default:
-                printf(">>>No such mode!\n");
-        }
-        getchar();
-        //可恶的换行符会进入default
-    }
-    return 0;
-}
 
 void showList(){
 	printf(">>>Begin\n");
@@ -85,7 +32,7 @@ bool insertItem(char name[],unsigned number){
 } 
 bool deleteItem(char name[],unsigned number){
     struct item* now=findItem(name);
-    if(now==NULL){
+    if(NULL==now){
         printf(">>>No such item\n");
         return 1;
     }
@@ -103,7 +50,7 @@ struct item* findItem(char name[]){
     }
     struct item* now=head;
     while(now){
-        if(strcmp(name,now->name)==0)return now;
+        if(0==strcmp(name,now->name))return now;
         now=now->next;
     }
     return NULL;
@@ -132,20 +79,4 @@ bool dumpList(){
     fclose(fp);
     printf(">>>Dumping End\n");
     return 0;
-}
-void readMe(){
-    printf("\n>>>>Read Me<<<<\n");
-    printf("Usage:\n");
-    printf("s : show the warehouse list\n");
-    printf("q : dump the list and quit\n");
-    printf("i X num : insert num Xs\n");
-    printf("d X num : delete num Xs\n\n");
-}
-inline FILE* fOpen(char * filename,char * mode){ 
-    FILE *temp=fopen(filename,mode);
-    if(temp)return temp;
-    else{ 
-        printf("\nFileOpen Error! %s : %s : %d\n",__FILE__,__func__,__LINE__);
-        return NULL;
-    }
 }
